grouped gemm: accept canonical (sum_m,k)/(l,n,k) layouts and flat SF buffers
The contiguous grouped GEMM SwiGLU/dSwiGLU wrappers and API classes now
additionally accept natural row-major inputs -- A (sum_m, k), B (l, n, k)
C-contiguous, dense C-contiguous SFA/SFB buffers (flat or physical atom
shape), prob (sum_m,) fp32/bf16 -- and normalize them internally. Canonical
SF buffers compile as flat 1-D pointers: the kernels already rebuild the
MMA-tiled SF layouts from the GEMM shapes and read only the base pointer.
Canonical calls return natural-shaped outputs; alpha defaults to cached
ones. Pre-permuted kernel-facing inputs keep working unchanged.

+++ python/cudnn/gemm/cutedsl/grouped/swiglu/grouped_gemm_swiglu_quant.py
@@ -2311,7 +2311,7 @@ class BlockScaledContiguousGroupedGemmKernel:
                 #
                 if cutlass.const_expr(prob is not None):
                     mPosition = tile_info[0] * self.mma_tiler[0] // cute.size(tiled_mma.thr_id.shape) + tidx
-                    mProb = prob[mPosition, 0, 0]
+                    mProb = prob[mPosition, 0, 0].to(cutlass.Float32)
                 else:
                     mProb = cutlass.Float32(1.0)
